+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Updated Azure Pipelines CI to use multi-stage pipeline to include publish step.
 - Fix error in `Get-PSAuthorizationString` when URI contains Query String
   parameters - fixes [Issue #14](https://github.com/PlagueHO/PSAuth/issues/14).
+- Added `OutFile` parameter to `Invoke-PSAuthRestMethod`.
 
 ## 0.1.3.51